[refactor] make srtm4 a python package (rather than a python module)
