+++ setup.py
@@ -34,7 +34,7 @@ requirements = ['filelock',
                 'numpy',
                 'requests'
                 ]
-                
+
 
 extras_require = {
         'test': ['pytest>=4.6', 'pytest-cov'],
@@ -49,7 +49,7 @@ setup(name="srtm4",
       url='https://github.com/cmla/srtm4',
       author='Carlo de Franchis, Enric Meinhardt, Gabriele Facciolo',
       author_email='carlo.de-franchis@ens-cachan.fr',
-      py_modules=['srtm4', 'srtm4c'],
+      packages=['srtm4'],
       install_requires=requirements,
       extras_require=extras_require,
       cmdclass={'develop': CustomDevelop,
